// src/commands.enum.ts
export enum COMMANDS {
  MOUSE_UP = "mouse_up",
  MOUSE_DOWN = "mouse_down",
  MOUSE_LEFT = "mouse_left",
  MOUSE_RIGHT = "mouse_right",
  MOUSE_POSITION = "mouse_position",
  DRAW_CIRCLE = "draw_circle",
  DRAW_RECT = "draw_rectangle",
  DRAW_SQUARE = "draw_square",
  PRINT_SCREEN = "prnt_scrn",
}